# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OPEN API CRM",
...
    "version": "v2"
  },
  "servers": [
    {
      "url": "https://crmconnect.misa.vn"
    }
  ],
  "paths": {
...
    "/api/v2/Contacts/id": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Lấy dữ liệu theo id",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "danh sách id",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": "danh sách id",
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/Contacts/code": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Lấy dữ liệu theo mã",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/Customers/id": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Lấy dữ liệu theo id",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "danh sách id",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": "danh sách id",
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/Customers/code": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "Lấy dữ liệu theo mã",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/Products": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Thực hiện thêm mới theo danh sách",
        "parameters": [
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Request body: Danh sách hàng hóa\r\n-\tXem thêm các trường có thể sử dụng tại ứng dụng CRM → Thiết lập → Dành cho nhà phát triển → API → tab Trường",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "owner_name": {
                    "type": "string",
                    "description": "Chủ sở hữu"
                  },
                  "product_code": {
                    "type": "string",
                    "description": "Mã hàng hóa"
                  },
                  "product_category": {
                    "type": "string",
                    "description": "Loại hàng hóa"
                  },
                  "usage_unit": {
                    "type": "string",
                    "description": "Đơn vị tính chính"
                  },
                  "product_name": {
                    "type": "string",
                    "description": "Tên hàng hóa"
                  },
                  "inactive": {
                    "type": "boolean",
                    "description": "Ngừng theo dõi"
                  },
                  "unit_price": {
                    "type": "string",
                    "description": "Đơn giá bán"
                  },
                  "tax": {
                    "type": "string",
                    "description": "Thuế GTGT"
                  },
                  "description": {
                    "type": "string",
                    "description": "Diễn giải"
                  },
                  "purchased_price": {
                    "type": "string",
                    "description": "Đơn giá mua"
                  },
                  "form_layout": {
                    "type": "string",
                    "description": "Bố cục"
                  },
                  "unit_price1": {
                    "type": "string",
                    "description": "Đơn giá bán 1"
                  },
                  "unit_price2": {
                    "type": "string",
                    "description": "Đơn giá bán 2"
                  },
                  "unit_price_fixed": {
                    "type": "string",
                    "description": "Đơn giá bán cố định"
                  },
                  "sale_description": {
                    "type": "string",
                    "description": "Diễn giải khi bán"
                  },
                  "product_properties": {
                    "type": "string",
                    "description": "Tính chất"
                  }
                }
              },
              "example": [
                {
                  "owner_name": "Lê Thị Lan (NV000001)",
                  "product_code": "HH00001",
                  "product_category": "Hàng hóa",
                  "usage_unit": "Bộ",
                  "product_name": "Tên hàng hóa",
                  "inactive": false,
                  "unit_price": "2000000",
                  "tax": "10%",
                  "description": "Mô tả",
                  "purchased_price": "1850000",
                  "form_layout": "Bố cục",
                  "unit_price1": "2100000",
                  "unit_price2": "2200000",
                  "unit_price_fixed": "2000000"
                }
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Thực hiện cập nhật theo danh sách",
        "parameters": [
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "requestBody": {
          "description": "Request body: Danh sách hàng hóa\r\n-\tXem thêm các trường có thể sử dụng tại ứng dụng CRM → Thiết lập → Dành cho nhà phát triển → API → tab Trường",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "owner_name": {
                    "type": "string",
                    "description": "Chủ sở hữu"
                  },
                  "product_code": {
                    "type": "string",
                    "description": "Mã hàng hóa"
                  },
                  "product_category": {
                    "type": "string",
                    "description": "Loại hàng hóa"
                  },
                  "usage_unit": {
                    "type": "string",
                    "description": "Đơn vị tính chính"
                  },
                  "product_name": {
                    "type": "string",
                    "description": "Tên hàng hóa"
                  },
                  "inactive": {
                    "type": "boolean",
                    "description": "Ngừng theo dõi"
                  },
                  "unit_price": {
                    "type": "string",
                    "description": "Đơn giá bán"
                  },
                  "tax": {
                    "type": "string",
                    "description": "Thuế GTGT"
                  },
                  "description": {
                    "type": "string",
                    "description": "Diễn giải"
                  },
                  "purchased_price": {
                    "type": "string",
                    "description": "Đơn giá mua"
                  },
                  "form_layout": {
                    "type": "string",
                    "description": "Bố cục"
                  },
                  "unit_price1": {
                    "type": "string",
                    "description": "Đơn giá bán 1"
                  },
                  "unit_price2": {
                    "type": "string",
                    "description": "Đơn giá bán 2"
                  },
                  "unit_price_fixed": {
                    "type": "string",
                    "description": "Đơn giá bán cố định"
                  },
                  "sale_description": {
                    "type": "string",
                    "description": "Diễn giải khi bán"
                  },
                  "product_properties": {
                    "type": "string",
                    "description": "Tính chất"
                  }
                }
              },
              "example": [
                {
                  "owner_name": "Lê Thị Lan (NV000001)",
                  "product_code": "HH00001",
                  "product_category": "Hàng hóa",
                  "usage_unit": "Bộ",
                  "product_name": "Tên hàng hóa",
                  "inactive": false,
                  "unit_price": "2000000",
                  "tax": "10%",
                  "description": "Mô tả",
                  "purchased_price": "1850000",
                  "form_layout": "Bố cục",
                  "unit_price1": "2100000",
                  "unit_price2": "2200000",
                  "unit_price_fixed": "2000000"
                }
              ]
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Thực hiện xoá theo danh sách id",
        "parameters": [
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                },
                "nullable": true
              }
            },
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                },
                "nullable": true
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                },
                "nullable": true
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "integer",
                  "format": "int64"
                },
                "nullable": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Paging lấy về danh sách có phân trang",
        "description": "page bắt đầu từ 0, pageSize max là 100, mặc định là 10, orderBy mặc định là theo modified_date desc",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Trang muốn xem",
            "schema": {
              "type": "integer",
              "description": "Trang muốn xem",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Số bản ghi trên 1 trang",
            "schema": {
              "type": "integer",
              "description": "Số bản ghi trên 1 trang",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "orderBy",
            "in": "query",
            "description": "Sắp xếp theo",
            "schema": {
              "type": "string",
              "description": "Sắp xếp theo",
              "default": "modified_date",
              "nullable": true
            }
          },
          {
            "name": "isDescending",
            "in": "query",
            "description": "Giảm dần",
            "schema": {
              "type": "boolean",
              "description": "Giảm dần",
              "default": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/Products/id": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Lấy dữ liệu theo id",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "danh sách id",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              },
              "description": "danh sách id",
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/Products/code": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Lấy dữ liệu theo mã",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseV2"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/Stocks/code": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "Lấy data theo mã Kho",
        "description": "Lấy data theo mã Kho",
        "parameters": [
          {
            "name": "stockCode",
            "in": "query",
            "description": "Mã Kho",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Mã Kho",
              "nullable": true
            }
          },
          {
            "name": "Clientid",
            "in": "header",
            "description": "Mã ứng dụng đăng ký trên crm",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockListApiResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ValidateInfoV2": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string",
            "nullable": true
          },
          "field_name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ValidateResultV2": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "description": "id bản ghi nếu thêm sửa thành công",
            "nullable": true
          },
          "validate_infos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidateInfoV2"
            },
            "description": "chi tiết lỗi của từng bản ghi",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApiResponseV2": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidateResultV2"
            },
            "description": "mỗi 1 object đem đi lưu sẽ tương ứng với 1 object này",
            "nullable": true
          },
          "data": {
            "description": "Dữ liệu trả về cho client",
            "nullable": true
          },
          "error_message": {
            "type": "string",
            "description": "thông tin lỗi",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "response trả về cho người dùng"
      },
...
      "StockListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "true: Request thực hiện thành công không có lỗi logic, ngược lại thì có lỗi"
          },
          "code": {
            "type": "integer",
            "description": "Thông tin mã lỗi",
            "format": "int32"
          },
          "dev_msg": {
            "type": "string",
            "description": "Thông báo lỗi cho nhà phát triển",
            "nullable": true
          },
          "user_msg": {
            "type": "string",
            "description": "Thông báo lỗi cho người dùng cuối",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Stock"
            },
            "description": "Dữ liệu trả về cho người dùng",
            "nullable": true
          },
          "validate_result": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidateResult"
            },
            "description": "Dữ liệu trả về cho người dùng",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Đối tượng response về client"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}